requirements.txt
absl-py==1.1.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beartype==0.18.5
blobfile==2.1.1
blosc2==2.0.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.10
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.13
docker-pycreds==0.4.0
# Local editable package from the original environment; install manually if this project needs it.
# drr-projector-function==0.0.0
easydict==1.11
einops==0.4.1
einops-exts==0.0.3
einx==0.3.0
ema-pytorch==0.0.8
executing==2.0.0
filelock==3.13.1
fonttools==4.34.4
frozendict==2.4.4
frozenlist==1.3.0
fsspec==2024.6.1
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.11.0
huggingface-hub==0.24.5
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.1.0
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.3
lpips==0.1.4
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
monai==0.9.0
mpi4py==3.1.5
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.2
nest-asyncio==1.5.8
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
ninja==1.11.1.1
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.8.1.78
packaging==23.2
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prefetch-generator==1.0.3
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.19.0
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.19.0
pyDeprecate==0.3.1
pydicom==2.3.0
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
pywin32==306
PyYAML==6.0
pyzmq==25.1.1
QtPy==2.4.1
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.1.5
rsa==4.8
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.3
sympy==1.13.1
tables==3.8.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.8.3
tokenizers==0.19.1
toml==0.10.2
torch==1.12.0+cu113
torch-tb-profiler==0.4.0
torchio==0.18.80
torchmetrics==0.9.1
torchstat==0.0.7
torchvision==0.13.0
tornado==6.3.3
tqdm==4.64.0
traitlets==5.11.2
transformers==4.44.0
typing_extensions==4.12.2
urllib3==1.26.16
ViTables==3.0.2
wandb==0.12.21
wcwidth==0.2.8
Werkzeug==2.1.2
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
